requirements.txt
google-cloud-language==2.4.1
Flask==2.0.3
